Added a new example, it shows how to manage sounds and music. Changed ChannelsCount to ChannelCount, since it makes more sense. Now PlayOffset and Duration in sounds and music is returned in seconds, as float.

+++ include/eepp/audio/csound.hpp
@@ -87,13 +87,13 @@ class EE_API cSound {
 		/** Get the Sound State */
 		Status State() const;
 
-		/** Get the current playing position of the sound */
-		virtual Uint32 PlayingOffset() const;
+		/** Get the current playing position of the sound in seconds */
+		virtual float PlayingOffset() const;
 
 		/** Set the current playing position of the sound
-		* @param TimeOffset : New playing position, expressed in miliseconds
+		* @param TimeOffset : New playing position, expressed in seconds
 		*/
-		virtual void PlayingOffset( const Uint32& TimeOffset );
+		virtual void PlayingOffset( const float& TimeOffset );
 
 		/** Assignment operator */
 		cSound& operator =(const cSound& Other);
